+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `RawPixel::CHANNELS` and `RawPixel::with_channel`
+- `F32x4` as a `RawPixel` implementation
+
+### Changed
+
+- Moved `crate::raw` to `crate::pixel::raw`
+- Renamed `U32x8888::new` to `U32x8888::from_u32`
+- Renamed `RawPixel::Value` to `RawPixel::Storage`
+- `RawPixel::set_channel` now returns `&mut Self`
+
 ## [0.3.0] - 2025-07-15
 
 ### Added